Refactor(app): useProfileImg 로 로직분리
